pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "substack2md"
dynamic = ["version"]
description = "Convert Substack posts to clean, Obsidian-friendly Markdown using your authenticated browser session."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    { name = "snapsynapse" }
]
keywords = ["substack", "markdown", "obsidian", "cdp", "chrome-devtools", "archival"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Utilities",
]
dependencies = [
    "websocket-client>=1.6.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "readability-lxml>=0.8.1",
    "markdownify>=0.11.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "responses>=0.24",
    "ruff>=0.4.0",
]

[project.urls]
Homepage = "https://github.com/snapsynapse/substack2md"
Repository = "https://github.com/snapsynapse/substack2md"
Issues = "https://github.com/snapsynapse/substack2md/issues"
Changelog = "https://github.com/snapsynapse/substack2md/blob/main/CHANGELOG.md"

[project.scripts]
substack2md = "substack2md.cli:main"

[tool.setuptools.packages.find]
include = ["substack2md*"]
exclude = ["tests*"]

[tool.setuptools.dynamic]
version = { attr = "substack2md._version.__version__" }

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
filterwarnings = [
    "ignore::DeprecationWarning:readability.*",
    "ignore::DeprecationWarning:bs4.*",
]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = [".venv", "build", "dist", ".claude"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "W",    # pycodestyle warnings
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # bugbear (common bug patterns)
]
ignore = [
    "E501",   # line too long -- trust the 100-char soft limit, let comments breathe
    "E741",   # ambiguous variable name (we use `l` and `m` in small scopes)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["F401", "F811"]  # pytest fixtures look unused to static analysis
"substack2md/_core.py" = ["E402"]  # dep-check runs before heavy imports by design

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
